Add `_apply_series_op` to Series as well.
